Serial I/O and string manipulation is working again
Well, it has been either three years or a couple of days now since I embarked on
the great divorce from the Arduino runtime. It uprooted a lot of stuff, namely all
I/O of any kind and all of the string code. So, now I just need to get digital GPIO
back up as well as the comparator and timers. Be back in another 5 years.